micro_companies/requirements.txt
aniso8601==9.0.1
bcrypt==4.0.1
blinker==1.6.2
certifi==2023.7.22
charset-normalizer==3.2.0
click==8.1.6
Faker==19.3.0
Flask==2.3.2
Flask-Cors==4.0.0
Flask-JWT-Extended==4.5.2
Flask-RESTful==0.3.10
Flask-SQLAlchemy==3.0.5
gevent==23.9.0
greenlet==2.0.2
gunicorn==21.2.0
honcho==1.1.0
idna==3.4
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.3
marshmallow==3.20.1
marshmallow-sqlalchemy==0.29.0
packaging==23.1
psycopg2-binary==2.9.7
PyJWT==2.8.0
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
six==1.16.0
SQLAlchemy==2.0.19
typing_extensions==4.7.1
urllib3==2.0.4
Werkzeug==2.3.7
zope.event==5.0
zope.interface==6.0
responses==0.23.3
coverage==7.1.0
responses==0.23.3